Unify binops further between Series and DataFrame.
